e2e tests for engine (#322)
