+++ e2e-tests/fixtures/engine/turbo-edits.md
@@ -1,0 +1,8 @@
+Example with turbo edit
+<dyad-edit path="foo/bar/file.js" description="turbo edit description">
+
+<!-- hello -->
+
+"making some edits"
+</dyad-edit>
+End of turbo edit
