+++ e2e-tests/snapshots/engine.spec.ts_send-message-to-engine-1.aria.yml
@@ -1,0 +1,4 @@
+- paragraph: "[dump] tc=turbo-edits"
+- paragraph: "[[dyad-dump-path=*]]"
+- button "Retry":
+  - img
